move ResumeWrap to ResumeWrap/ResumeWrap
