+++ src/elements/ResumeWrap/ResumeWrap.js
@@ -3,12 +3,12 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import styles from './index.css';
+import styles from './ResumeWrap.css';
 import { InputWrap } from 'elements/InputWrap/index';
 import { ListWrap } from 'elements/Section/ListWrap';
+import { Button } from 'elements/Button/index';
 import { getRedactedFileName, getFileName } from 'lib/resume';
-import { HelpTextWrap } from 'elements/Section/HelpTextWrap/index';
-import { editCvName } from 'store/actions/cv';
+import { editCvName, removeCv } from 'store/actions/cv';
 
 class PreResumeWrap extends Component {
 	constructor(props) {
@@ -25,11 +25,12 @@ class PreResumeWrap extends Component {
 		this.setState({
 			name: ev.target.value,
 		});
-		this.props.editCvName(this.path, ev.target.value);
+		this.props.editCvName(this.props.path, ev.target.value);
 	}
 
 	render() {
 		const { fileName, name, redactedFileName } = this.state;
+		const { path } = this.props;
 
 		return (
 			<div>
@@ -44,11 +45,15 @@ class PreResumeWrap extends Component {
 							onChange={e => this.onChange(e)}
 						/>
 					</InputWrap>
+					<Button
+						secondary
+						onClick={() => {
+							this.props.removeCv(path);
+						}}
+					>
+						Remove from list
+					</Button>
 				</ListWrap>
-				<HelpTextWrap>
-					A new file, {redactedFileName} will be saved alongside the original,
-					blocking out the name you provided as well as emails and urls
-				</HelpTextWrap>
 			</div>
 		);
 	}
@@ -58,5 +63,6 @@ export const ResumeWrap = connect(
 	state => ({}),
 	dispatch => ({
 		editCvName: bindActionCreators(editCvName, dispatch),
+		removeCv: bindActionCreators(removeCv, dispatch),
 	})
 )(PreResumeWrap);
